+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * stax pull-request - will honor base branch replacement setting
  * stax commit - will come back to original branch even if user haven't used Pull Request creation flag
  * stax about - new command with brief information about stax, link to the website and basic license infrmation
+ * git checkout -> git switch - replaced all internal invocations of git checkout to git switch
 
 0.10.3
  * stax rebase - now ceates folder structure for rebase file